Emit the message from the right script.
